backend/requirements.txt
django
djangorestframework
psycopg2
redis
celery
pandas
django-filter
django-cors-headers
Pillow